handling uint to int

+++ src/core/filter.go
@@ -44,7 +44,7 @@ func GetReleases(action string, gitHubToken string, days *uint) (map[string]inte
 	}
 
 	now := time.Now()
-	interval := time.Duration(*days * dayInNanos)
+	interval := time.Duration(*days * uint(dayInNanos))
 	limit := now.Add(-interval)
 
 	url := apiBaseURL + action + "/releases"
